+++ src/gamspy/_algebra/expression.py
@@ -10,6 +10,7 @@ import gamspy._algebra.operable as operable
 import gamspy._algebra.operation as operation
 import gamspy._validation as validation
 import gamspy.utils as utils
+from gamspy._config import get_option
 from gamspy._extrinsic import ExtrinsicFunction
 from gamspy._symbols.implicits import ImplicitSet
 from gamspy._symbols.implicits.implicit_symbol import ImplicitSymbol
@@ -24,6 +25,17 @@ if TYPE_CHECKING:
 
 GMS_MAX_LINE_LENGTH = 80000
 LINE_LENGTH_OFFSET = 79000
+
+LATEX_DATA_MAP = {
+    "=g=": "\\geq",
+    "=l=": "\\leq",
+    "=e=": "=",
+    "*": "\\cdot",
+    "and": "\\wedge",
+    "or": "\\vee",
+    "xor": "\\oplus",
+    "$": "|",
+}
 
 
 @dataclass
@@ -83,7 +95,10 @@ class Expression(operable.Operable):
         if data == "=" and isinstance(right, Expression):
             right._fix_equalities()
 
-        self.representation = self._create_output_str()
+        if get_option("LAZY_EVALUATION"):
+            self._representation = None
+        else:
+            self._representation = self._create_output_str()
         self.where = condition.Condition(self)
         self._create_domain()
         left_control = getattr(left, "controlled_domain", [])
@@ -92,10 +107,17 @@ class Expression(operable.Operable):
             {*left_control, *right_control}
         )
         self.container = None
-        if left is not None and hasattr(left, "container"):
-            self.container = left.container
-        elif right is not None and hasattr(right, "container"):
-            self.container = right.container
+        if hasattr(left, "container"):
+            self.container = left.container  # type: ignore
+        elif hasattr(right, "container"):
+            self.container = right.container  # type: ignore
+
+    @property
+    def representation(self) -> str:
+        if self._representation is None:
+            self._representation = self._create_output_str()
+
+        return self._representation
 
     def _create_domain(self):
         for loc, result in [
@@ -239,7 +261,9 @@ class Expression(operable.Operable):
 
     def _replace_operator(self, operator: str):
         self.data = operator
-        self.representation = self._create_output_str()
+
+        if not get_option("LAZY_EVALUATION"):
+            self._representation = self._create_output_str()
 
     def latexRepr(self) -> str:
         """
@@ -356,7 +380,7 @@ class Expression(operable.Operable):
                     stack.append(root.right)
 
                 stack.append(root)
-                root = root.left if hasattr(root, "left") else None
+                root = root.left if hasattr(root, "left") else None  # type: ignore
 
             if len(stack) == 0:
                 break
@@ -367,7 +391,8 @@ class Expression(operable.Operable):
                 if root.data in EQ_MAP:
                     root._replace_operator(EQ_MAP[root.data])
                 else:
-                    root.representation = root._create_output_str()
+                    if not get_option("LAZY_EVALUATION"):
+                        root._representation = root._create_output_str()
 
             last_item = peek(stack)
             if (
@@ -473,6 +498,9 @@ class Expression(operable.Operable):
     def _validate_definition(
         self, control_stack: list[Set | Alias | ImplicitSet]
     ) -> None:
+        if not get_option("DOMAIN_VALIDATION"):
+            return
+
         stack = []
 
         node = self.right
@@ -564,4 +592,5 @@ class SetExpression(Expression):
                         f"Incompatible operand `{self.left}` for the set operation `{self.data}`."
                     )
 
-        self.representation = self._create_output_str()
+        if not get_option("LAZY_EVALUATION"):
+            self._representation = self._create_output_str()