+++ src/gamspy/_types.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import builtins
 from typing import Union
 
 from typing_extensions import TypeAlias
@@ -8,4 +7,4 @@ from typing_extensions import TypeAlias
 from gamspy._algebra.operable import Operable
 
 OperableType: TypeAlias = Union[Operable, int, float]
-EllipsisType = builtins.ellipsis
+EllipsisType: TypeAlias = type(...)  # type: ignore